requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
cffi==1.16.0 ; python_version >= '3.8'
cryptography==41.0.5
django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0 ; python_version >= '3.6'
djangorestframework-simplejwt==5.3.0
fernet==1.0.1
pillow==10.1.0
pyaes==1.6.1
pycparser==2.21
pyjwt==2.8.0 ; python_version >= '3.7'
pytz==2023.3.post1
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.8.0 ; python_version < '3.11'
